requirements.txt
openai>=1.55.0         # Para interações via OpenRouter
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
httpx>=0.27.2
redis>=5.2.0
python-dotenv>=1.0.1
numpy>=2.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
